Warm_up_Challenges/Jumping_on_the_Clouds/Solution_py/Solution.py
#!/bin/python3

import math
import os
import random
import re
import sys

#
# Complete the 'jumpingOnClouds' function below.
#
# The function is expected to return an INTEGER.
# The function accepts INTEGER_ARRAY c as parameter.
#


def jumpingOnClouds(c):
    # Write your code here
    count = 0
    route = []
    for cloud in c:
        if cloud == 0:
            count += 1
            route.append(cloud)
        else:
            route = []
        if len(route) == 3:
            count -= 1
            route = [0]
    return count - 1

if __name__ == '__main__':
    fptr = open(os.environ['OUTPUT_PATH'], 'w')

    n = int(input().strip())

    c = list(map(int, input().rstrip().split()))

    result = jumpingOnClouds(c)

    fptr.write(str(result) + '\n')

    fptr.close()
